Added CYBER and Updated Excel Data Import
Added cyber building code to program and added feature to check if an asset has a value before checking if its blank to ensure it does not try to reference a null variable and throw an exception. In the future may look at removing the check for null data inside of checking for if there is that data present using the boolean variable hasVariableName as this seems redundant.