Excel Compatibility Update
Added the functionality of being able to create assets without certain fields being present in Excel.
Fixed a bug where data fields' boolean values for if they were written to were always true. It now checks to confirm the value is not None from excel and if it is, it does not set that value on the asset.